Move deconstruction semantic tests out of the code gen tests assy
