+++ src/Compilers/CSharp/Test/Semantic/CSharpCompilerSemanticTest.csproj
@@ -75,6 +75,7 @@
     <Compile Include="FlowAnalysis\TryLockUsingStatementTests.cs" />
     <Compile Include="Semantics\BindingAsyncTasklikeMoreTests.cs" />
     <Compile Include="Semantics\BindingAsyncTasklikeTests.cs" />
+    <Compile Include="Semantics\DeconstructionTests.cs" />
     <Compile Include="Semantics\ImportsTests.cs" />
     <Compile Include="Semantics\AccessCheckTests.cs" />
     <Compile Include="Semantics\AccessibilityTests.cs" />
